backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
diff-match-patch==20241021
Django==4.2.30
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.1
django-import-export==4.4.0
django-ratelimit==4.1.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework-api-key==3.1.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
exceptiongroup==1.3.1
frozenlist==1.8.0
humanize==4.13.0
idna==3.13
inflection==0.5.1
kombu==5.6.2
multidict==6.7.1
packaging==26.1
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.12
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.0.1
reportlab==4.4.10
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
twilio==9.10.5
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
yarl==1.22.0
faker
unidecode
